# Return a categorical DataFrame with Sampling