### Your checklist for this pull request
Please review the [guidelines for contributing](CONTRIBUTING.md) to this repository.

- [x] Make sure you are requesting to **pull a topic/feature/bugfix branch** (right side). Don't request your master!
- [x] Make sure you are making a pull request against the **dev branch** (left side). Also you should start *your branch* off *our dev*.
- [x] Check the commit's or even all commits' message styles matches our requested structure.

### List of changes to the codebase in this pull request
- Update the `_return_samples` utility function to return a pandas DataFrame with `category` dtypes, as opposed to string. Using `category` has a lower memory impact, and makes it faster to compute the state counts (3x faster on 100k samples of Alarm). Update the `_return_samples` to use `state_names` instead of `state_names_map`.

#### Before (604f1f9)
```python
In [1]: from pgmpy.utils import get_example_model

In [2]: from pgmpy.sampling import BayesianModelSampling

In [3]: model = get_example_model('alarm')

In [4]: samples = BayesianModelSampling(model).forward_sample(size=100_000)

In [5]: samples.info()
<class 'pandas.core.frame.DataFrame'>
RangeIndex: 100000 entries, 0 to 99999
Data columns (total 37 columns):
 #   Column        Non-Null Count   Dtype
---  ------        --------------   -----
 0   MINVOLSET     100000 non-null  object
 1   VENTMACH      100000 non-null  object
 ...
 35  PCWP          100000 non-null  object
 36  CVP           100000 non-null  object
dtypes: object(37)
memory usage: 28.2+ MB

In [6]: from pgmpy.estimators import BaseEstimator

In [7]: estimator = BaseEstimator(data=samples)

In [8]: %timeit -n1 -r1 estimator.state_counts('CATECHOL', model.predecessors('CATECHOL'))
212 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

#### After
```python
In [1]: from pgmpy.utils import get_example_model

In [2]: from pgmpy.sampling import BayesianModelSampling

In [3]: model = get_example_model('alarm')

In [4]: samples = BayesianModelSampling(model).forward_sample(size=100_000)

In [5]: samples.info()
<class 'pandas.core.frame.DataFrame'>
RangeIndex: 100000 entries, 0 to 99999
Data columns (total 37 columns):
 #   Column        Non-Null Count   Dtype
---  ------        --------------   -----
 0   MINVOLSET     100000 non-null  category
 1   VENTMACH      100000 non-null  category
 ...
 35  PCWP          100000 non-null  category
 36  CVP           100000 non-null  category
dtypes: category(37)
memory usage: 3.5 MB

In [6]: from pgmpy.estimators import BaseEstimator

In [7]: estimator = BaseEstimator(data=samples)

In [8]: %timeit -n1 -r1 estimator.state_counts('CATECHOL', model.predecessors('CATECHOL'))
72.7 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```