@ankurankan 2020-08-11
@tristandeleu I wasn't aware of the pandas had a Categorical datatype. Thanks for the PR.